Group Markers based on a raster on the map.
Create one group marker where multiple markers are in close vicinity.
The raster is fixed (width and height in px - 30px in this commit)
and the markers are all regrouped each time the zoom level is changed.

The group marker's position is set to the average of all contained markers.